backend/requirements.txt
uvicorn
fastapi
pydantic-settings
redis
hiredis
asyncpg
google-genai
httpx